+++ server-new/src/api/routers/__init__.py
@@ -3,9 +3,11 @@
 from .email_configs import router as email_configs_router
 from .pdf_files import router as pdf_files_router
 from .pdf_templates import router as pdf_templates_router
+from .pipelines import router as pipelines_router
 
 __all__ = [
     'email_configs_router',
     'pdf_files_router',
     'pdf_templates_router',
+    'pipelines_router',
 ]